requirements.txt
amqp==5.3.1
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
chirpstack-api==4.10.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.21.1
Flask==3.1.0
Flask-Cors==5.0.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
idna==3.10
influxdb-client==1.48.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kombu==5.4.2
libsass==0.23.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.24.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
reactivex==4.0.4
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
setuptools==75.6.0
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
tenacity==9.0.0
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.9
